chore: split tests according to contracts
